* rename histo to histogram * add documentation for image export * move validation to `Ingredient.py`
